+++ Bencodex/IOffloadOptions.cs
@@ -9,11 +9,14 @@ namespace Bencodex
     {
         /// <summary>
         /// Determines whether a given <paramref name="indirectValue"/> should be embedded or
-        /// offloaded.
+        /// offloaded.  <em>This method must be deterministic.</em>
         /// </summary>
         /// <param name="indirectValue">A value to determine whether to embed or offload.</param>
         /// <returns><see langword="true"/> if <paramref name="indirectValue"/> should be embedded;
         /// <see langword="false"/> if it should be offloaded.</returns>
+        /// <remarks>Note that returning <see langword="true"/> does not mean the entire value
+        /// including its subvalues is embedded, but only the container is.  Subvalues in it
+        /// are separately determined using distinct calls on this method.</remarks>
         public bool Embeds(in IndirectValue indirectValue);
 
         /// <summary>